refactor: redux builder module
