+++ libs/frontend/model/infra/api/src/GraphqlOperationOptions.ts
@@ -3,5 +3,5 @@ import type { RequestInit } from 'graphql-request/dist/types.dom'
 export interface GraphqlOperationOptions<
   TVariables extends Record<string, any> = Record<string, any>,
 > extends Pick<RequestInit, 'headers'> {
-  variables?: TVariables
+  variables?: TVariables | null
 }
