+++ libs/frontend/model/infra/api/src/client.ts
@@ -4,9 +4,9 @@ import type { RequestInit } from 'graphql-request/dist/types.dom'
 const API_ROUTE = '/api/graphql'
 let client: GraphQLClient | undefined
 
-export const getGraphQLClient = (options?: RequestInit) => {
+export const getGraphQLClient = (options?: RequestInit | null) => {
   if (!client) {
-    client = new GraphQLClient(API_ROUTE, options)
+    client = new GraphQLClient(API_ROUTE, options ?? undefined)
   }
 
   return client
