+++ libs/frontend/modules/builder/src/index.ts
@@ -1,10 +1,11 @@
 export * from './Builder'
-export * from './containers/builderState'
-export * from './containers/useBuilderHandlers'
 export * from './main-pane'
 export * from './meta-pane'
 export * from './overlay-toolbar'
 export * from './PageBuilderTemplate'
 export * from './renderer'
+export * from './store/builderState'
+export * from './store/useBuilderHoverHandlers'
+export * from './store/useOnRendered'
 export * from './templates/BuilderSidebarNavigation'
 export * from './templates/BuilderTemplate'
