+++ libs/frontend/modules/builder/src/main-pane/MainPaneBuilder.tsx
@@ -2,155 +2,57 @@ import { IElement } from '@codelab/frontend/abstract/core'
 import {
   CreateElementModal,
   useElementGraphContext,
-  useMoveElementMutation,
 } from '@codelab/frontend/modules/element'
-import { EntityType, useCrudModalForm } from '@codelab/frontend/view/components'
 import {
   MainPaneTemplate,
   MainPaneTemplateProps,
 } from '@codelab/frontend/view/templates'
-import { ElementTree } from '@codelab/shared/core'
 import { Dropdown, Tree as AntdTree } from 'antd'
-import { TreeProps } from 'antd/lib/tree'
+import { DataNode } from 'rc-tree/lib/interface'
 import React, { useState } from 'react'
-import { useHotkeys } from 'react-hotkeys-hook'
+import { useDispatch } from 'react-redux'
 import tw from 'twin.macro'
-import { useBuilderSelection } from '../containers/builderState'
+import { builderActions } from '../store/builderState'
+import { useBuilderSelectedElement } from '../store/useBuilderSelectedElement'
 import { ElementContextMenu } from './ElementContextMenu'
+import { useElementTreeDrop } from './useElementTreeDrop'
 import { useExpandedNodes } from './useExpandedNodes'
 
 export type MainPaneBuilderTemplateProps = MainPaneTemplateProps
 
 /**
- * Reusable builder pane, which renders an Antd tree for visualizing the DOM tree
+ * Base builder pane, which renders an Antd tree for visualizing the DOM tree
+ * Requires ElementGraphContext
  */
 export const MainPaneBuilder = ({
   children,
   ...props
 }: MainPaneBuilderTemplateProps) => {
-  const { elementTree, elementId } = useElementGraphContext()
-  const tree = elementTree ?? new ElementTree({ edges: [], vertices: [] })
+  const { elementTree } = useElementGraphContext()
+  const { selectedElement } = useBuilderSelectedElement()
+  const { setExpandedNodeIds, expandedNodeIds } = useExpandedNodes(elementTree)
+  const antdTree = elementTree.getAntdTree()
+  const { isMoving, handleDrop } = useElementTreeDrop(elementTree)
+  const dispatch = useDispatch()
+  const resetSelection = () => dispatch(builderActions.resetSelection)
 
-  const {
-    setHoveringElement,
-    setSelectedElement,
-    resetSelection,
-    state: { selectedElement },
-  } = useBuilderSelection()
+  const setHoveringElement = (elementId?: string) =>
+    dispatch(builderActions.hoverElement({ elementId }))
 
-  const { openDeleteModal } = useCrudModalForm(EntityType.Element)
-  useHotkeys(
-    'del,backspace',
-    () => {
-      if (selectedElement) {
-        openDeleteModal([selectedElement.id], selectedElement)
-      }
-    },
-    { enabled: !!selectedElement },
-    [selectedElement],
-  )
-  useHotkeys(
-    'esc',
-    () => {
-      if (selectedElement) {
-        setSelectedElement(undefined)
-      }
-    },
-    { enabled: !!selectedElement },
-    [selectedElement],
-  )
-
-  const { setExpandedNodeIds, expandedNodeIds } = useExpandedNodes(
-    tree,
-    selectedElement,
-  )
-
-  const [contextMenuItemId, setContextMenuNodeId] = useState<string | null>(
-    null,
-  )
-
-  const antdTree = tree.getAntdTree()
-
-  const [moveElement, { isLoading: isLoadingMoveElement }] =
-    useMoveElementMutation()
-
-  const handleDrop: TreeProps['onDrop'] = (e) => {
-    // This can be optimized to be handled in the API
-    // It is also buggy, because it doesn't handle the case where the two nodes have the same order
-
-    const dragNodeId = (e.dragNode as any).id
-    const dropNodeId = (e.node as any).id
-
-    if (e.dropToGap) {
-      // Switch spots with the element next to the drop indicator
-
-      const dropNodeParentId = tree.getParentOf(dropNodeId)?.id
-      const dropElementOrder = tree.getOrderInParent(dropNodeId)
-      const originalDragElementOrder = tree.getOrderInParent(dragNodeId)
-
-      if (dropNodeParentId) {
-        moveElement({
-          variables: {
-            input: {
-              elementId: dragNodeId,
-              moveData: {
-                parentElementId: dropNodeParentId,
-                order:
-                  dropElementOrder === originalDragElementOrder
-                    ? dropElementOrder + 1
-                    : dropElementOrder,
-              },
-            },
-          },
-        }).catch(console.error)
-
-        moveElement({
-          variables: {
-            input: {
-              elementId: dropNodeId,
-              moveData: {
-                parentElementId: dropNodeParentId,
-                order: originalDragElementOrder,
-              },
-            },
-          },
-        }).catch(console.error)
-      }
-    } else {
-      // FIXME
-      // Move the dragged element as a child to the dropped element
-      // This is buggy, since e.dropPosition does not match our ordering system
-      // it causes issues when moving elements up
-      return moveElement({
-        variables: {
-          input: {
-            elementId: dragNodeId,
-            moveData: {
-              parentElementId: dropNodeId,
-              order: e.dropPosition,
-            },
-          },
-        },
-      })
-    }
-
-    return void 0
-  }
+  const setSelectedElement = (elementId?: string) =>
+    dispatch(builderActions.selectElement({ elementId }))
 
   return (
     <MainPaneTemplate
       {...props}
       containerProps={{
         style: { width: '100%', height: '100%' },
-        onClick: () => {
-          setContextMenuNodeId(null)
-        },
       }}
     >
       <div onClick={() => resetSelection()}>
-        {tree ? (
+        {elementTree ? (
           <AntdTree
-            disabled={isLoadingMoveElement}
+            disabled={isMoving}
             className="draggable-tree"
             blockNode
             expandedKeys={expandedNodeIds}
@@ -158,62 +60,14 @@ export const MainPaneBuilder = ({
             onExpand={(expandedKeys) => setExpandedNodeIds(expandedKeys)}
             onDrop={handleDrop}
             selectedKeys={selectedElement ? [selectedElement.id] : []}
-            onMouseEnter={({ node: dataNode }) => {
-              const element = tree.getVertex((dataNode as any).id?.toString())
-
-              setHoveringElement(element)
-            }}
+            onMouseEnter={({ node: { id } }: any) => setHoveringElement(id)}
+            onMouseLeave={() => setHoveringElement(undefined)}
             onClick={(e) => e.stopPropagation()}
-            onMouseLeave={() => {
-              setHoveringElement(undefined)
-            }}
             onSelect={([id], { nativeEvent }) => {
               nativeEvent.stopPropagation()
-
-              const element = tree.getVertex(id?.toString())
-              setSelectedElement(element)
+              setSelectedElement(id?.toString())
             }}
-            titleRender={(node) => {
-              const element = node as any as IElement
-              const label = element.name
-              const nodeId = element.id
-              const atomName = element.atom?.name || element.atom?.type
-
-              return (
-                <div data-cy={`atom-${label}`}>
-                  <Dropdown
-                    onVisibleChange={() => setContextMenuNodeId(nodeId)}
-                    visible={contextMenuItemId === nodeId}
-                    overlay={
-                      <>
-                        <div
-                          css={tw`fixed inset-0`}
-                          onClick={(e) => {
-                            setContextMenuNodeId(null)
-                            e.stopPropagation()
-                          }}
-                        />
-                        <ElementContextMenu
-                          // We need to manually hide the context menu, otherwise it stays open
-                          onClick={() => setContextMenuNodeId(null)}
-                          element={node as any}
-                        />
-                      </>
-                    }
-                    trigger={['contextMenu']}
-                  >
-                    <div>
-                      {label}{' '}
-                      {atomName && (
-                        <span css={tw`text-gray-400 text-xs`}>
-                          ({atomName})
-                        </span>
-                      )}
-                    </div>
-                  </Dropdown>
-                </div>
-              )
-            }}
+            titleRender={(node) => <TreeItemTitle node={node} />}
             treeData={antdTree ? [antdTree] : undefined}
           />
         ) : null}
@@ -225,5 +79,48 @@ export const MainPaneBuilder = ({
         {children}
       </div>
     </MainPaneTemplate>
+  )
+}
+
+const TreeItemTitle = ({ node }: { node: DataNode }) => {
+  const [contextMenuItemId, setContextMenuNodeId] = useState<string | null>(
+    null,
+  )
+
+  const element = node as any as IElement
+  const { name: label, id: nodeId, atom } = element
+  const atomName = atom?.name || atom?.type
+
+  return (
+    <div data-cy={`atom-${label}`}>
+      <Dropdown
+        onVisibleChange={() => setContextMenuNodeId(nodeId)}
+        visible={contextMenuItemId === nodeId}
+        overlay={
+          <>
+            <div
+              css={tw`fixed inset-0`}
+              onClick={(e) => {
+                setContextMenuNodeId(null)
+                e.stopPropagation()
+              }}
+            />
+            <ElementContextMenu
+              // We need to manually hide the context menu, otherwise it stays open
+              onClick={() => setContextMenuNodeId(null)}
+              element={node as any}
+            />
+          </>
+        }
+        trigger={['contextMenu']}
+      >
+        <div>
+          {label}{' '}
+          {atomName && (
+            <span css={tw`text-gray-400 text-xs`}>({atomName})</span>
+          )}
+        </div>
+      </Dropdown>
+    </div>
   )
 }
