+++ libs/frontend/modules/builder/src/main-pane/useExpandedNodes.tsx
@@ -1,22 +1,21 @@
-import { IElement } from '@codelab/frontend/abstract/core'
 import { ElementTree } from '@codelab/shared/core'
 import { useEffect, useState } from 'react'
+import { useBuilderSelectedElement } from '../store/useBuilderSelectedElement'
 
-export const useExpandedNodes = (
-  tree: ElementTree,
-  selectedPageElement?: IElement,
-) => {
+export const useExpandedNodes = (tree: ElementTree) => {
+  const { selectedElementId } = useBuilderSelectedElement()
+
   const [expandedNodeIds, setExpandedNodeIds] = useState<
     Array<string | number>
   >([])
 
   // When we select a element, expand all tree nodes from the root to the selected elements
   useEffect(() => {
-    if (!selectedPageElement) {
+    if (!selectedElementId) {
       return
     }
 
-    const pathResult = tree.getPathFromRoot(selectedPageElement.id)
+    const pathResult = tree.getPathFromRoot(selectedElementId)
 
     // If there is a path (there should always be, it's a tree after all), go through each node
     // of the path and keep track of all nodes that need to get expanded
@@ -30,7 +29,7 @@ export const useExpandedNodes = (
 
       return [...prevState, ...toExpand]
     })
-  }, [tree, selectedPageElement])
+  }, [tree, selectedElementId])
 
   return { expandedNodeIds, setExpandedNodeIds }
 }
