+++ libs/frontend/modules/builder/src/meta-pane/PropsInspectorTab.tsx
@@ -1,82 +1,22 @@
-import {
-  useGetElementQuery,
-  useUpdateElementPropsMutation,
-} from '@codelab/frontend/modules/element'
-import { notify } from '@codelab/frontend/shared/utils'
 import { MonacoEditor } from '@codelab/frontend/view/components'
-import { propSafeStringify } from '@codelab/shared/utils'
 import Button from 'antd/lib/button'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import tw from 'twin.macro'
-import { useBuilder } from '../containers/builderState'
+import { usePropsInspector } from './usePropsInspector'
 
 export interface ElementPropsSectionProps {
   elementId: string
 }
 
 const PropsInspectorTab = ({ elementId }: ElementPropsSectionProps) => {
-  const [persistedPropsValue, setPersistedPropsValue] = useState<
-    string | undefined
-  >()
-
-  const { data } = useGetElementQuery({
-    variables: { input: { where: { id: elementId } } },
-  })
-
   const {
-    state: { currentElementProps },
+    save,
+    lastRenderedPropsString,
+    persistedProps,
+    setPersistedProps,
+    isLoading,
     setExtraPropsForElement,
-  } = useBuilder()
-
-  const [mutate, { isLoading }] = useUpdateElementPropsMutation({})
-  const element = data?.getElement
-
-  useEffect(() => {
-    if (element?.props) {
-      try {
-        setPersistedPropsValue(
-          JSON.stringify(JSON.parse(element.props), null, 4),
-        )
-      } catch (e) {
-        console.warn("Couldn't parse element props", element?.props)
-      }
-    }
-  }, [element?.props])
-
-  useEffect(() => {
-    return () => {
-      setExtraPropsForElement(elementId, {})
-    }
-  }, [elementId, setExtraPropsForElement])
-
-  if (!element) {
-    return null
-  }
-
-  const save = async () => {
-    if (!persistedPropsValue) {
-      notify({ title: 'Invalid json', type: 'warning' })
-
-      return
-    }
-
-    try {
-      await mutate({
-        variables: {
-          input: {
-            elementId,
-            props: JSON.stringify(JSON.parse(persistedPropsValue)),
-          },
-        },
-      })
-    } catch (e) {
-      notify({ title: 'Invalid json', type: 'warning' })
-    }
-  }
-
-  const currentPropsValue = propSafeStringify(
-    currentElementProps[elementId] ?? {},
-  )
+  } = usePropsInspector(elementId)
 
   return (
     <div css={tw`grid grid-cols-2 gap-x-8 h-full`}>
@@ -85,7 +25,7 @@ const PropsInspectorTab = ({ elementId }: ElementPropsSectionProps) => {
         <MonacoEditor
           containerProps={{ style: { height: '100%' } }}
           editorOptions={{ language: 'json', readOnly: true }}
-          value={currentPropsValue}
+          value={lastRenderedPropsString}
         />
       </div>
 
@@ -99,12 +39,12 @@ const PropsInspectorTab = ({ elementId }: ElementPropsSectionProps) => {
         <MonacoEditor
           containerProps={{ style: { height: '100%' } }}
           editorOptions={{ language: 'json' }}
-          value={persistedPropsValue}
+          value={persistedProps}
           onChange={(v) => {
-            setPersistedPropsValue(v)
+            setPersistedProps(v)
 
             try {
-              setExtraPropsForElement(elementId, JSON.parse(v))
+              setExtraPropsForElement(JSON.parse(v))
             } catch (e) {
               //
             }
