+++ libs/frontend/modules/builder/src/renderer/pipes/extraPropsPipe.ts
@@ -1,0 +1,14 @@
+import { mergeProps } from '@codelab/shared/utils'
+import { RenderPipeFactory } from '../types/RenderPipe'
+
+/**
+ * Adds context.extraProps
+ */
+export const extraPropsPipe: RenderPipeFactory =
+  (next) => (element, context, props) => {
+    return next(
+      element,
+      context,
+      context.extraProps ? mergeProps(props, context.extraProps) : props,
+    )
+  }
