+++ libs/frontend/modules/builder/src/renderer/renderFactory.tsx
@@ -1,13 +1,13 @@
 import { RenderNode } from '@codelab/frontend/abstract/core'
 import { RenderContext } from '@codelab/frontend/presenter/container'
 import React from 'react'
-import { renderElement } from './handlers/renderElement'
+import { renderPipeline } from './renderPipeline'
 
 /**
  * Creates a React Component from a {@link RenderNode}
  */
 export const renderFactory = (node: RenderNode, context: RenderContext) => {
-  const toRender: React.ReactNode = renderElement(node, context)
+  const toRender: React.ReactNode = renderPipeline(node, context, {})
 
   return Array.isArray(toRender) && toRender?.length ? toRender[0] : toRender
 }
