--- libs/frontend/modules/builder/src/renderer/types/RenderHandler.ts
@@ -1,10 +1,0 @@
-import { RenderNode } from '@codelab/frontend/abstract/core'
-import {
-  RenderContext,
-  RenderOutput,
-} from '@codelab/frontend/presenter/container'
-
-export type RenderHandler = (
-  node: RenderNode,
-  metadata: RenderContext,
-) => RenderOutput
