+++ libs/frontend/modules/builder/src/store/redux.d.ts
@@ -1,0 +1,7 @@
+import { BuilderState } from './builderState'
+
+declare module 'react-redux' {
+  interface DefaultRootState {
+    builder: BuilderState
+  }
+}
