+++ libs/shared/abstract/core/src/props/props.interface.ts
@@ -1,1 +1,3 @@
 export type IProps = Record<string, any>
+
+export type PropsByElementId = Record<string, IProps | undefined>
