+++ package.json
@@ -317,6 +317,7 @@
     "eslint-import-resolver-node": "0.3.6",
     "eslint-import-resolver-typescript": "2.5.0",
     "eslint-plugin-cypress": "2.12.1",
+    "eslint-plugin-deprecate-import": "^2.0.0",
     "eslint-plugin-graphql": "4.0.0",
     "eslint-plugin-import": "2.25.2",
     "eslint-plugin-jest-formatting": "3.0.0",
